deployment/requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
Pillow
plotly
scikit-learn==1.3.0
tensorflow==2.15.0